requirements.txt
PyInstaller
pyqt5-stubs
pyserial
pywin32
WMI
